Refactor infer.py into separate components
