+++ runner/app/python/infer/streamer/__init__.py
@@ -1,0 +1,3 @@
+from .streamer import PipelineStreamer
+
+__all__ = ["PipelineStreamer"]
